use FindBin to search for library modules.
Traditionally Perl would alwayws look for modules in the current directory which was ok if you executed these scripts from the directory they are in. Due to potential vulnerabilties, the inclusion of CWD (or ".") in @INC is now frowned upon, and has been removed from Perl 5.26 on. Using FindBin ensures that the .pm files are looked for in the same directory as the script itself, which should have been done long ago as it also allows calling the scripts from different directories.